requirements.txt
numpy==1.24.4
pandas==1.5.3
python-dotenv==1.0.1
scikit_learn==1.6.1
together==1.3.11
tqdm==4.66.5
